requirements.txt
# GPU Cluster Multi-View Reconstruction – Python Dependencies
# Core API + pipeline dependencies (GPU-specific packages like torch,
# dust3r, mast3r, trellis are installed separately via setup scripts)

# API framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Data validation
pydantic>=2.7.0

# Redis
redis>=5.0.0

# Image processing
Pillow>=10.3.0
numpy>=1.26.0

# Background removal (subject isolation)
rembg>=2.0.50
onnxruntime-gpu>=1.18.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

